requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-generativeai==0.8.3
grpcio==1.68.0
grpcio-status==1.68.0
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
pillow==11.0.0
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
